# core/aggregator/api/swagger.yaml
openapi: 3.0.0
info:
  title: Swagger Gandalf
  description: "This is a sample Petstore server.  You can find \nout more about Swagger\
    \ at \n[http://swagger.io](http://swagger.io) or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n"
  termsOfService: http://swagger.io/terms/
  contact:
    email: romain.fairant@orness.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0-oas3
...
servers:
- url: https://virtserver.swaggerhub.com/ditrit/Gandalf/1.0.0
- url: http://virtserver.swaggerhub.com/ditrit/Gandalf/1.0.0
...
paths:
...
  /domain:
    get:
      tags:
      - domain
      summary: List domains
      operationId: listDomain
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200_1'
                x-content-type: application/json
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /domain/{domainId}:
    get:
      tags:
      - domain
      summary: Find domain by ID
      description: Returns a single domain
      operationId: getDomainById
      parameters:
      - name: domainId
        in: path
        description: ID of domain to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        "400":
          description: Invalid ID supplied
        "404":
          description: Domain not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    put:
      tags:
      - domain
      summary: Updates a domain
      operationId: updateDomain
      parameters:
      - name: domainId
        in: path
        description: ID of domain that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateDomain'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        "400":
          description: Invalid ID supplied | Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    delete:
      tags:
      - domain
      summary: Deletes a domain
      operationId: deleteDomain
      parameters:
      - name: domainId
        in: path
        description: Domain id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Succes
        "400":
          description: Invalid ID supplied
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /domain/{domainName}:
    get:
      tags:
      - domain
      summary: Find domain by Name
      description: Returns a single Domain
      operationId: getDomainByName
      parameters:
      - name: domainName
        in: path
        description: Name of domain to return
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        "404":
          description: Domain not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    post:
      tags:
      - domain
      summary: Create a domain
      operationId: createDomain
      parameters:
      - name: domainName
        in: path
        description: Name of parent domain
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateDomain'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        "400":
          description: Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
...
  /eventType:
    get:
      tags:
      - eventType
      summary: List eventTypes
      operationId: listEventType
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200_7'
                x-content-type: application/json
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    post:
      tags:
      - eventType
      summary: Create a eventType
      operationId: createEventType
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateEventType'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        "400":
          description: Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /eventType/{eventTypeId}:
    get:
      tags:
      - eventType
      summary: Find eventType by ID
      description: Returns a single eventType
      operationId: getEventTypeById
      parameters:
      - name: eventTypeId
        in: path
        description: ID of role to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        "400":
          description: Invalid ID supplied
        "404":
          description: EventType not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    put:
      tags:
      - eventType
      summary: Updates a eventType
      operationId: updateEventType
      parameters:
      - name: eventTypeId
        in: path
        description: ID of eventType that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateEventType'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        "400":
          description: Invalid ID supplied | Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    delete:
      tags:
      - eventType
      summary: Deletes a eventType
      operationId: deleteEventType
      parameters:
      - name: api_key
        in: header
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: eventTypeId
        in: path
        description: EventType id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Success
        "400":
          description: Invalid ID supplied
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /eventType/{eventTypeName}:
    get:
      tags:
      - eventType
      summary: Find eventType by Name
      description: Returns a single EventType
      operationId: getEventTypeByName
      parameters:
      - name: eventTypeName
        in: path
        description: Name of eventType to return
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        "404":
          description: EventType not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
...
components:
  schemas:
    Model:
      type: object
      properties:
        id:
          type: integer
          description: The ID.
        createdAt:
          type: string
          description: The created date.
          format: date-time
        updatedAt:
          type: string
          description: The updated date.
          format: date-time
        deletedAt:
          type: string
          description: The deleted date.
          format: date-time
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        deletedAt: 2000-01-23T04:56:07.000+00:00
        id: 0
        updatedAt: 2000-01-23T04:56:07.000+00:00
...
    Domain:
      type: object
      properties:
        model:
          $ref: '#/components/schemas/Model'
        name:
          type: string
          description: The domain name.
      example:
        name: name
        model:
          createdAt: 2000-01-23T04:56:07.000+00:00
          deletedAt: 2000-01-23T04:56:07.000+00:00
          id: 0
          updatedAt: 2000-01-23T04:56:07.000+00:00
...
    EventType:
      type: object
      properties:
        model:
          $ref: '#/components/schemas/Model'
        name:
          type: string
          description: The eventType name.
        schema:
          type: string
        pivotId:
          type: integer
          description: The pivot id.
        productConnectorId:
          type: integer
          description: The productConnector id.
      example:
        schema: schema
        productConnectorId: 6
        pivotId: 0
        name: name
        model:
          createdAt: 2000-01-23T04:56:07.000+00:00
          deletedAt: 2000-01-23T04:56:07.000+00:00
          id: 0
          updatedAt: 2000-01-23T04:56:07.000+00:00
...
    inline_response_200_1:
      type: object
      properties:
        schema:
          $ref: '#/components/schemas/Domain'
      example:
        schema:
          name: name
          model:
            createdAt: 2000-01-23T04:56:07.000+00:00
            deletedAt: 2000-01-23T04:56:07.000+00:00
            id: 0
            updatedAt: 2000-01-23T04:56:07.000+00:00
...
    inline_response_200_7:
      type: object
      properties:
        schema:
          $ref: '#/components/schemas/EventType'
      example:
        schema:
          schema: schema
          productConnectorId: 6
          pivotId: 0
          name: name
          model:
            createdAt: 2000-01-23T04:56:07.000+00:00
            deletedAt: 2000-01-23T04:56:07.000+00:00
            id: 0
            updatedAt: 2000-01-23T04:56:07.000+00:00
...
  requestBodies:
...
    createOrUpdateDomain:
      description: The domain to create.
      content:
        application/json:
          schema:
            required:
            - name
            type: object
            properties:
              name:
                type: string
...
    createOrUpdateEventType:
      description: The eventType to create.
      content:
        application/json:
          schema:
            required:
            - name
            - pivotID
            - productConnectorID
            - schema
            type: object
            properties:
              name:
                type: string
              schema:
                type: string
              pivotID:
                type: integer
              productConnectorID:
                type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT